coreected the flow and separate the page class
